update: 例外処理を追加

+++ docker/download/download.sh
@@ -27,9 +27,13 @@ gguf_name=${hf_model_file_name}_${convert_outtype}.gguf
 
 # convert model
 cd /data/download
-if [ ! -e /data/models/${gguf_name} ];then 
-    python3 /opt/llama.cpp/convert.py ${hf_model_file_name}/ --outfile ${gguf_name} --outtype ${convert_outtype}
-    mv ${gguf_name} /data/models
+if [ -e /data/download/${hf_model_file_name} ];then
+    if [ ! -e /data/models/${gguf_name} ];then 
+        python3 /opt/llama.cpp/convert.py ${hf_model_file_name}/ --outfile ${gguf_name} --outtype ${convert_outtype}
+        mv ${gguf_name} /data/models
+    else
+        echo "[log] file already exists: /data/models/${gguf_name}"
+    fi
 else
-    echo "[log] file already exists: /data/models/${gguf_name}"
+    echo "[error] directory not exists: /data/download/${hf_model_file_name}"
 fi
